+++ awscfncli/commands/changeset/execute.py
@@ -3,7 +3,6 @@
 __author__ = 'kotaimen'
 __date__ = '13/01/2017'
 
-import boto3
 import click
 
 from ..utils import boto3_exception_handler, pretty_print_config, \
@@ -37,17 +36,22 @@ def execute(ctx, config_file, changeset_name, no_wait):
     CONFIG_FILE         Stack configuration file.
     CHANGESET_NAME      The name of the change set.
     """
+    session = ctx.obj['session']
+
     # load config
     stack_config = load_stack_config(config_file)
     pretty_print_config(stack_config)
-    load_template_body(stack_config)
+    load_template_body(session, stack_config)
 
     # connect to cfn
     region = stack_config.pop('Region')
 
+    # remove unused parameters
+    stack_config.pop('Package', None)
+
     # execute change set
-    client = boto3.client('cloudformation', region_name=region)
-    cfn = boto3.resource('cloudformation', region_name=region)
+    client = session.client('cloudformation', region_name=region)
+    cfn = session.resource('cloudformation', region_name=region)
     stack = cfn.Stack(stack_config['StackName'])
 
     if stack.stack_status == 'REVIEW_IN_PROGRESS':
@@ -66,7 +70,7 @@ def execute(ctx, config_file, changeset_name, no_wait):
         return
 
     # start event tailing
-    start_tail_stack_events_daemon(stack, latest_events=5)
+    start_tail_stack_events_daemon(session, stack, latest_events=5)
 
     # wait until update complete
     waiter = client.get_waiter(waiter_model)