+++ awscfncli/commands/stack/deploy.py
@@ -3,7 +3,6 @@
 __author__ = 'kotaimen'
 __date__ = '11/01/2017'
 
-import boto3
 import click
 
 from ...cli import stack
@@ -44,12 +43,15 @@ def deploy(ctx, config_file, no_wait, on_failure, canned_policy):
 
     CONFIG_FILE         Stack configuration file.
     """
+
+    session = ctx.obj['session']
+
     # load config
     stack_config = load_stack_config(config_file)
     click.echo('Deploying stack...')
     pretty_print_config(stack_config)
 
-    load_template_body(stack_config)
+    load_template_body(session, stack_config)
 
     # option handling
     if on_failure is not None:
@@ -62,7 +64,11 @@ def deploy(ctx, config_file, no_wait, on_failure, canned_policy):
 
     # connect to cfn
     region = stack_config.pop('Region')
-    cfn = boto3.resource('cloudformation', region_name=region)
+
+    # remove unused parameters
+    stack_config.pop('Package', None)
+
+    cfn = session.resource('cloudformation', region_name=region)
 
     # create stack
     stack = cfn.create_stack(**stack_config)
@@ -74,10 +80,10 @@ def deploy(ctx, config_file, no_wait, on_failure, canned_policy):
         return
 
     # start event tailing
-    start_tail_stack_events_daemon(stack, latest_events=0)
+    start_tail_stack_events_daemon(session, stack, latest_events=0)
 
     # wait until update complete
-    waiter = boto3.client('cloudformation', region_name=region).get_waiter(
+    waiter = session.client('cloudformation', region_name=region).get_waiter(
         'stack_create_complete')
     waiter.wait(StackName=stack_id)